requirements.txt
scipy
numpy
tqdm
openai
matplotlib
python-dotenv
requests
nest-asyncio
pykan
scikit-learn
PyYAML
torch
pandas
stopit
setuptools
jax
jaxlib